+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+## 0.2.0
+- Migrate to TypeScript
+- Add support for any code or style language
+- Use vscode language assosiated to the file instead of file extension
+
 ## 0.1.0
 - Initial release